refactor mutation logic to hooks
