+++ apps/web/app/components/session/annotations/AnnotationContent.tsx
@@ -1,111 +1,18 @@
-import { useState } from "react";
 import { Button } from "~/components/Button";
 import { Icon } from "~/components/ui/icon";
 import { AnnotationCard } from "./AnnotationCard";
-
-const mockAnnotations = [
-  {
-    id: 1,
-    title: "Annotation Title Herer",
-    timestamp: "00:00 - 01:01",
-    author: "Garrett Lo",
-    date: "Sep 25",
-    description: "",
-  },
-  {
-    id: 2,
-    title: "Annotation Title Here",
-    timestamp: "00:00 - 01:01",
-    author: "Garrett Lo",
-    date: "Sep 25, 2024",
-    description: "",
-  },
-  {
-    id: 3,
-    title: "Annotation Title Here",
-    timestamp: "00:00 - 01:01",
-    author: "Garrett Lo",
-    date: "Sep 25",
-    description: "Written text description.",
-  },
-  {
-    id: 4,
-    title: "Annotation Title Here",
-    timestamp: "00:00 - 01:01",
-    author: "Garrett Lo",
-    date: "Sep 25",
-    description:
-      "Written text description notes this one shows the overflow wrapping with multiple displayed lines.",
-  },
-  {
-    id: 5,
-    title: "Annotation Title Here",
-    timestamp: "00:00 - 01:01",
-    author: "Garrett Lo",
-    date: "Sep 25",
-    description:
-      "Written text description notes showing wrapping text with new line implementation.\nThis is what it looks like when they press enter for new line.",
-  },
-  {
-    id: 6,
-    title: "Annotation Title Here",
-    timestamp: "00:00 - 01:01",
-    author: "Garrett Lo",
-    date: "Sep 25",
-    description:
-      "Written text description notes showing wrapping text with new line implementation.\nThis is what it looks like when they press enter for new line.",
-  },
-  {
-    id: 7,
-    title: "Annotation Title Here",
-    timestamp: "00:00 - 01:01",
-    author: "Garrett Lo",
-    date: "Sep 25",
-    description:
-      "Written text description notes showing wrapping text with new line implementation.\nThis is what it looks like when they press enter for new line.",
-  },
-];
+import { useAnnotations } from "~/hooks/useAnnotations";
 
 export function AnnotationContent() {
-  const [annotations, setAnnotations] = useState(mockAnnotations);
-  const [selectedId, setSelectedId] = useState<number | null>(null);
-
-  const handleSelect = (id: number) => setSelectedId(id);
-  const handleDeselect = () => setSelectedId(null);
-
-  const handleTitleChange = (id: number, next: string) => {
-    setAnnotations((list) =>
-      list.map((a) => (a.id === id ? { ...a, title: next } : a))
-    );
-  };
-
-  const handleDescriptionChange = (id: number, next: string) => {
-    setAnnotations((list) =>
-      list.map((a) => (a.id === id ? { ...a, description: next } : a))
-    );
-  };
-
-  const handleNewAnnotation = () => {
-    const nextId = annotations.length
-      ? Math.max(...annotations.map((a) => a.id)) + 1
-      : 1;
-    const newItem = {
-      id: nextId,
-      title: "",
-      timestamp: "00:00 - 00:00",
-      author: "Garrett Lo",
-      date: `${new Date()
-        .toLocaleDateString("en-US", {
-          month: "short",
-          day: "numeric",
-          year: "numeric",
-        })
-        .replace(",", "")}`,
-      description: "",
-    };
-    setAnnotations([newItem, ...annotations]);
-    setSelectedId(nextId);
-  };
+  const {
+    annotations,
+    selectedId,
+    handleSelect,
+    handleDeselect,
+    handleTitleChange,
+    handleDescriptionChange,
+    handleNewAnnotation,
+  } = useAnnotations();
 
   const effectiveSelectedId = selectedId ?? null;
